# documentation/api/openapi/carehub-openapi.yaml
openapi: 3.1.0
info:
  title: Carehub API
  description: "API per gestione pazienti, medici, appuntamenti"
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer-jwt: []
paths:
...
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
  /api/auth/register-user:
    post:
      tags:
      - auth-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
  /api/auth/register-patient:
    post:
      tags:
      - auth-controller
      operationId: registerPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SimpleResult"
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/reports/analytics/patient/{patientId}/overview:
    get:
      tags:
      - reports-analytics-controller
      operationId: overviewByPatient
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AnalyticsOverviewDto"
  /api/reports/analytics/patient/{patientId}/monthly:
    get:
      tags:
      - reports-analytics-controller
      operationId: monthlyByPatient
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenueDto"
  /api/reports/analytics/overview:
    get:
      tags:
      - reports-analytics-controller
      operationId: overviewAll
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AnalyticsOverviewDto"
  /api/reports/analytics/monthly:
    get:
      tags:
      - reports-analytics-controller
      operationId: monthlyAll
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenueDto"
  /api/reports/analytics/doctor/{doctorId}/overview:
    get:
      tags:
      - reports-analytics-controller
      operationId: overviewByDoctor
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AnalyticsOverviewDto"
  /api/reports/analytics/doctor/{doctorId}/monthly:
    get:
      tags:
      - reports-analytics-controller
      operationId: monthlyByDoctor
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenueDto"
  /api/reports/analytics/by-doctor:
    get:
      tags:
      - reports-analytics-controller
      operationId: byDoctor
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: patientId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DoctorRevenueDto"
...
components:
  schemas:
...
    SimpleResult:
      type: object
      properties:
        result:
          type: string
        resultMessageHeader:
          type: string
        resultMessage:
          type: string
...
    RegisterUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        fiscalCode:
          type: string
        birthDate:
          type: string
          format: date
        specialization:
          type: string
        roles:
          type: array
          items:
            type: string
          uniqueItems: true
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
...
    AnalyticsOverviewDto:
      type: object
      properties:
        totalRevenue:
          type: number
        reportsCount:
          type: integer
          format: int64
        averageCost:
          type: number
    MonthlyRevenueDto:
      type: object
      properties:
        month:
          type: string
        totalRevenue:
          type: number
        reportsCount:
          type: integer
          format: int64
    DoctorRevenueDto:
      type: object
      properties:
        doctorId:
          type: integer
          format: int64
        totalRevenue:
          type: number
        reportsCount:
          type: integer
          format: int64
...
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT